pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vocabulary-learning"
version = "0.1.0"
authors = [
    { name="Aurelien Nioche", email="nioche.aurelien@gmail.com" },
]
description = "A tool for learning Japanese vocabulary"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "firebase-admin>=6.2.0",
    "python-dotenv>=1.0.0",
    "rich>=13.6.0",
    "pykakasi>=2.2.1",
    "Levenshtein>=0.21.1",
    "httpx>=0.25.0",
    "deep-translator>=1.11.4",
]

[project.urls]
"Homepage" = "https://github.com/AurelienNioche/vocabulary-learning"
"Bug Tracker" = "https://github.com/AurelienNioche/vocabulary-learning/issues"

[project.scripts]
vocab = "vocabulary_learning.main:main"
fix-progress = "vocabulary_learning.utils.fix_progress:fix_progress"
fix-duplicates = "vocabulary_learning.utils.fix_duplicates:fix_duplicates"
sync-to-firebase = "vocabulary_learning.utils.sync_to_firebase:sync_to_firebase"
fix-translations = "vocabulary_learning.utils.fix_translations:fix_translations"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:httpx.*:",
    "ignore::UserWarning"
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.hatch.build.targets.wheel]
packages = ["vocabulary_learning"]

[tool.hatch.build.targets.wheel.sources]
"vocabulary_learning" = "vocabulary_learning"
"vocabulary_learning/default_data" = "vocabulary_learning/default_data"
